Extract classes to individual files
