+++ Empire/LogLevel.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Empire;
+
+class LogLevel {
+    const WARNING   = 'warning';
+    const INFO      = 'info';
+    const DEBUG     = 'debug';
+}
